refactor(draw-cap): move cap views in a dedicated file
